// src/app/modules/preventivatore/preventivatore-dynamic/services/content/content-interface.ts
export interface ContentInterface {
  header?: {
    data: any;
  };
  hero?: any;
  bg_img_hero?: any;
  how_works?: any;
  what_to_know?: any;
  more_info?: any;
  for_who?: any;
  contact_us?: any;

  // generic containers
  container_1?: any;
  container_2?: any;
  container_3?: any;
  container_4?: any;
  container_5?: any;
  container_6?: any;
  container_7?: any;
  container_8?: any;
}